requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
langchain-huggingface
faiss-cpu
langchain-google-genai
streamlit